% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ilogit}
\alias{ilogit}
\title{\code{ilogit}  Calculate the inverse logit of x}
\usage{
ilogit(x)
}
\arguments{
\item{x}{the log-odds between -inf and +inf}
}
\value{
the proportion associated with x
}
\description{
\code{ilogit}  Calculate the inverse logit of x
}
\examples{
ilogit(0)
}

