% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ilogit}
\alias{ilogit}
\title{\code{ilogit}  Calculate the inverse logit of w}
\usage{
ilogit(x)
}
\arguments{
\item{w}{the log-odds between -inf and +inf}
}
\value{
the proportion associated with w
}
\description{
Calculate the proportion from a log-odds
}
\details{
exp(w)/(1+exp(w))
}
\examples{
ilogit(0)
}

