% Generated by roxygen2 (4.0.1): do not edit by hand
\name{logit}
\alias{logit}
\title{Calculate the log odds from a proportion}
\usage{
logit(x)
}
\arguments{
\item{x}{Proportion.  Between 0 and 1.}
}
\value{
the log of the odds associated with x
}
\description{
This should be the details.  It might be longer than the description
Or, at least that is what I am planning
}
\details{
Actually the above was description, this should be details

What if this is multiple lines
}
\examples{
logit(.5)
}

